Split operators like add, sub into modules
